add schema structure
