+++ packages/superadmin/components/products/schema/productSchema/index.ts
@@ -1,0 +1,4 @@
+import { productSchema } from "./createProduct.Schema";
+import { searchProductSchema } from "./searchProduct.Schema";
+
+export {productSchema , searchProductSchema};
